src/main/java/com/project/springboot/plantidati/model/CalendarEntryId.java
package com.project.springboot.plantidati.model;

import jakarta.persistence.Embeddable;
import java.io.Serializable;
import java.util.Objects;

@Embeddable
public class CalendarEntryId implements Serializable {

    private String calendarId;

    private String entryId;

    // Getters and setters
    public String getCalendarId() {
        return calendarId;
    }

    public void setCalendarId(String calendarId) {
        this.calendarId = calendarId;
    }

    public String getEntryId() {
        return entryId;
    }

    public void setEntryId(String entryId) {
        this.entryId = entryId;
    }

    // Equals and hashCode methods
    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        CalendarEntryId that = (CalendarEntryId) o;
        return Objects.equals(calendarId, that.calendarId) && Objects.equals(entryId, that.entryId);
    }

    @Override
    public int hashCode() {
        return Objects.hash(calendarId, entryId);
    }
}